use MockSpanFactory in ReactRouterRoutingProvider unit test suite
